+++ apps/next-playground/components/pages/TestPage.tsx
@@ -1,5 +1,17 @@
-const TestPage = () => {
-  return <div>TestPage</div>;
-};
+import { Link as AsheeLink } from "asheeui";
+import NextLink from "next/link";
 
-export default TestPage;
+export default function TestPage() {
+  return (
+    <AsheeLink
+      href="/dashboard"
+      linkComponent={NextLink}
+      linkProps={{ prefetch: true }}
+      variant="default"
+      color="primary"
+      size="md"
+      underline="hover">
+      Dashboard
+    </AsheeLink>
+  );
+}